requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
chardet==5.2.0
Django
django-pyodbc==1.1.3
et_xmlfile==2.0.0
mssql-django==1.5
numpy
openpyxl==3.1.5
pandas
Pillow==10.0.0
pyodbc==4.0.39
python-bidi==0.6.3
python-dateutil==2.9.0.post0
pytz==2024.2
reportlab==4.2.5
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
google-cloud-storage